Add ability to download playlists from Youtube Music

+++ home/autodotfiles/functions/youtube-dl.tag.zsh
@@ -3,12 +3,19 @@
 # Function meant for internal use
 # Tags a file downloaded by 'ytdlalbum' with additional metadata
 
+write_id4cover=1
+
+if [[ $1 = "--no-id4cover" ]]; then
+    write_id4cover=0
+    shift
+fi
+
 if [[ ! -f "$1" ]]; then
     echo "\e[91mFile not found: \e[95m$1"
     return 1
 fi
 
-if [[ ! "$(basename "$1")" =~ ^([0-9]+)\.([0-9]+)\.([a-zA-Z0-9_\-]+)\.(.*)$ ]]; then
+if [[ ! "$(basename "$1")" =~ ^([0-9]+)\.([0-9NA]+)\.([a-zA-Z0-9_\-]+)\.(.*)$ ]]; then
     echo "\e[91mInvalid filename format: \e[95m$1"
     return 1
 fi
@@ -22,6 +29,12 @@ fi
 
 out="$dir/${match[1]}.${match[4]:t:r}.$out_ext"
 
+if [[ ${match[2]} != "NA" ]]; then
+    year="${match[2]}"
+else
+    year=""
+fi
+
 echo "[ADF-Tagger] Tagging \e[95m$out\e[0m..."
 
 ffmpeg -hide_banner -loglevel error \
@@ -32,13 +45,13 @@ ffmpeg -hide_banner -loglevel error \
     -metadata DESCRIPTION="" \
     -metadata description="" \
     -metadata TRACK="$((match[1]))" \
-    -metadata year="$((match[2]))" \
-    -metadata YEAR="$((match[2]))" \
-    -metadata date="$((match[2]))" \
-    -metadata DATE="$((match[2]))" \
+    -metadata year="$year" \
+    -metadata YEAR="$year" \
+    -metadata date="$year" \
+    -metadata DATE="$year" \
     "$out"
 
-if [[ ! -f "$dir/__id4cover.txt" ]]; then
+if (( $write_id4cover )) && [[ ! -f "$dir/__id4cover.txt" ]]; then
     echo "${match[3]}" > "$dir/__id4cover.txt"
 fi
